rename exx_symmetry as module_exx_symmetry
