+++ source/module_ri/module_exx_symmetry/irreducible_sector.h
