+++ source/module_ri/module_exx_symmetry/symmetry_rotation.cpp
