+++ source/module_ri/module_exx_symmetry/test/CMakeLists.txt
