+++ source/module_ri/module_exx_symmetry/test/symmetry_rotation_test.cpp
